Add DistrictsTestCaseMixin, and rename *TestCaseMixins to just *Mixins and move them to their own module
