+++ test_project/test_app/tests/test_manage_command.py
@@ -1,99 +1,53 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 
-import re
-
 from unittest import skipIf
 
 from django import VERSION as django_version
 from django.test import TestCase, override_settings
 from django.core.management import call_command
 
-from cities.models import (Country, Region, Subregion, City, PostalCode,
-                           AlternativeName)
+from cities.models import (Country, Region, Subregion, City, District,
+                           PostalCode, AlternativeName)
 
-
-class NoInvalidSlugsTestCaseMixin(object):
-    def test_no_invalid_slugs(self):
-        self.assertEqual(Country.objects.filter(slug__startswith='invalid').count(), 0)
-        self.assertEqual(Region.objects.filter(slug__startswith='invalid').count(), 0)
-        self.assertEqual(Subregion.objects.filter(slug__startswith='invalid').count(), 0)
-        self.assertEqual(City.objects.filter(slug__startswith='invalid').count(), 0)
-        self.assertEqual(PostalCode.objects.filter(slug__startswith='invalid').count(), 0)
-
-
-class CountriesTestCaseMixin(object):
-    def test_num_countries(self):
-        self.assertEqual(Country.objects.all().count(), self.num_countries)
-
-
-class RegionsTestCaseMixin(object):
-    def test_num_regions(self):
-        self.assertEqual(Region.objects.count(), self.num_regions)
-
-    def test_num_ad_regions(self):
-        self.assertEqual(
-            Region.objects.filter(country__code='AD').count(),
-            self.num_ad_regions)
-
-    def test_num_ua_regions(self):
-        self.assertEqual(
-            Region.objects.filter(country__code='UA').count(),
-            self.num_ua_regions)
-
-
-class SubregionsTestCaseMixin(object):
-    def test_num_subregions(self):
-        self.assertEqual(Subregion.objects.count(), self.num_subregions)
-
-
-class CitiesTestCaseMixin(object):
-    def test_num_cities(self):
-        self.assertEqual(City.objects.count(), self.num_cities)
-
-    def test_num_ua_cities(self):
-        self.assertEqual(
-            City.objects.filter(region__country__code='UA').count(),
-            self.num_ua_cities)
-
-
-class AlternativeNameTestCaseMixin(object):
-    def test_num_alternative_names(self):
-        self.assertEqual(AlternativeName.objects.count(), self.num_alt_names)
-
-    def test_num_not_und_alternative_names(self):
-        self.assertEqual(
-            AlternativeName.objects.exclude(language_code='und').count(),
-            self.num_not_und_alt_names)
-
-
-class PostalCodeTestCaseMixin(object):
-    def test_num_postal_codes(self):
-        self.assertEqual(PostalCode.objects.count(), self.num_postal_codes)
-
-    def test_postal_code_slugs(self):
-        pc = PostalCode.objects.get(country__code='RU', code='102104')
-        self.assertEqual(pc.code, '102104')
-        self.assertTrue(len(pc.slug) <= 255)
-
-        slug_rgx = re.compile(r'\d+-102104', re.UNICODE)
-
-        slug = PostalCode.objects.get(country__code='RU', code='102104').slug
-
-        # The unittest module in Python 2 does not have an assertRegex
-        if hasattr(self, 'assertRegex'):
-            self.assertRegex(slug, slug_rgx)
-        else:
-            m = slug_rgx.match(slug)
-
-            self.assertIsNotNone(m)
-            self.assertEqual(m.group(0)[-7:], '-102104')
+from ..mixins import (
+    NoInvalidSlugsMixin, CountriesMixin, RegionsMixin, SubregionsMixin,
+    CitiesMixin, DistrictsMixin, AlternativeNamesMixin, PostalCodesMixin)
 
 
 @override_settings(CITIES_IGNORE_EMPTY_REGIONS=True)
 class IgnoreEmptyRegionManageCommandTestCase(
-        NoInvalidSlugsTestCaseMixin, CountriesTestCaseMixin,
-        RegionsTestCaseMixin, SubregionsTestCaseMixin, CitiesTestCaseMixin,
+        NoInvalidSlugsMixin, CountriesMixin, RegionsMixin, SubregionsMixin,
+        CitiesMixin, DistrictsMixin, TestCase):
+    num_countries = 250
+    num_regions = 170
+    num_ad_regions = 7
+    num_ua_regions = 27
+    num_subregions = 4926
+    num_cities = 114
+    num_ua_cities = 50
+    num_districts = 0
+
+    @classmethod
+    def setUpTestData(cls):
+        # Run the import command only once
+        super(IgnoreEmptyRegionManageCommandTestCase, cls).setUpTestData()
+        call_command('cities', force=True, **{
+            'import': 'country,region,subregion,city,district',
+        })
+        cls.counts = {
+            'countries': Country.objects.count(),
+            'regions': Region.objects.count(),
+            'subregions': Subregion.objects.count(),
+            'cities': City.objects.count(),
+            'districts': District.objects.count(),
+            'postal_codes': PostalCode.objects.count(),
+        }
+
+
+class ManageCommandTestCase(
+        NoInvalidSlugsMixin, CountriesMixin, RegionsMixin, SubregionsMixin,
+        CitiesMixin, DistrictsMixin, AlternativeNamesMixin, PostalCodesMixin,
         TestCase):
     num_countries = 250
     num_regions = 170
@@ -102,34 +56,7 @@ class IgnoreEmptyRegionManageCommandTestCase(
     num_subregions = 4926
     num_cities = 114
     num_ua_cities = 50
-
-    @classmethod
-    def setUpTestData(cls):
-        # Run the import command only once
-        super(IgnoreEmptyRegionManageCommandTestCase, cls).setUpTestData()
-        call_command('cities', force=True, **{
-            'import': 'country,region,subregion,city',
-        })
-        cls.counts = {
-            'countries': Country.objects.count(),
-            'regions': Region.objects.count(),
-            'subregions': Subregion.objects.count(),
-            'cities': City.objects.count(),
-            'postal_codes': PostalCode.objects.count(),
-        }
-
-
-class ManageCommandTestCase(
-        NoInvalidSlugsTestCaseMixin, CountriesTestCaseMixin,
-        RegionsTestCaseMixin, SubregionsTestCaseMixin, CitiesTestCaseMixin,
-        AlternativeNameTestCaseMixin, PostalCodeTestCaseMixin, TestCase):
-    num_countries = 250
-    num_regions = 170
-    num_ad_regions = 7
-    num_ua_regions = 27
-    num_subregions = 4926
-    num_cities = 114
-    num_ua_cities = 50
+    num_districts = 0
     num_alt_names = 2945
     num_not_und_alt_names = 579
     num_postal_codes = 13
@@ -139,13 +66,14 @@ class ManageCommandTestCase(
         # Run the import command only once
         super(ManageCommandTestCase, cls).setUpTestData()
         call_command('cities', force=True, **{
-            'import': 'country,region,subregion,city,alt_name,postal_code',
+            'import': 'country,region,subregion,city,district,alt_name,postal_code',
         })
         cls.counts = {
             'countries': Country.objects.count(),
             'regions': Region.objects.count(),
             'subregions': Subregion.objects.count(),
             'cities': City.objects.count(),
+            'districts': District.objects.count(),
             'alt_names': AlternativeName.objects.count(),
             'postal_codes': PostalCode.objects.count(),
         }
@@ -163,6 +91,7 @@ class ManageCommandTestCase(
         self.assertEqual(Region.objects.count(), self.counts['regions'])
         self.assertEqual(Subregion.objects.count(), self.counts['subregions'])
         self.assertEqual(City.objects.count(), self.counts['cities'])
+        self.assertEqual(District.objects.count(), self.counts['districts'])
         self.assertEqual(AlternativeName.objects.count(), self.counts['alt_names'])
         self.assertEqual(PostalCode.objects.count(), self.counts['postal_codes'])
 
@@ -171,9 +100,8 @@ class ManageCommandTestCase(
 @skipIf(django_version < (1, 8), "Django < 1.8, skipping test with CITIES_LOCALES=['all']")
 @override_settings(CITIES_LOCALES=['all'])
 class AllLocalesManageCommandTestCase(
-        NoInvalidSlugsTestCaseMixin, CountriesTestCaseMixin,
-        RegionsTestCaseMixin, SubregionsTestCaseMixin, CitiesTestCaseMixin,
-        AlternativeNameTestCaseMixin, TestCase):
+        NoInvalidSlugsMixin, CountriesMixin, RegionsMixin, SubregionsMixin,
+        CitiesMixin, DistrictsMixin, AlternativeNamesMixin, TestCase):
     num_countries = 250
     num_regions = 170
     num_ad_regions = 7
@@ -181,6 +109,7 @@ class AllLocalesManageCommandTestCase(
     num_subregions = 4926
     num_cities = 114
     num_ua_cities = 50
+    num_districts = 0
     num_alt_names = 7760
     num_not_und_alt_names = 5394
 
@@ -189,12 +118,13 @@ class AllLocalesManageCommandTestCase(
         # Run the import command only once
         super(AllLocalesManageCommandTestCase, cls).setUpTestData()
         call_command('cities', force=True, **{
-            'import': 'country,region,subregion,city,alt_name',
+            'import': 'country,region,subregion,city,district,alt_name',
         })
         cls.counts = {
             'countries': Country.objects.count(),
             'regions': Region.objects.count(),
             'subregions': Subregion.objects.count(),
             'cities': City.objects.count(),
+            'districts': District.objects.count(),
             'alt_names': AlternativeName.objects.count(),
         }
